+++ Chapter.Net.WPF.Controls/Bases/StackPanelBase.cs
@@ -1,0 +1,17 @@
+﻿// -----------------------------------------------------------------------------------------------------------------
+// <copyright file="StackPanelBase.cs" company="my-libraries">
+//     Copyright (c) David Wendland. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------------------------------------------------
+
+using System.Windows.Controls;
+
+namespace Chapter.Net.WPF.Controls.Bases
+{
+    /// <summary>
+    ///     Base class for stack panels.
+    /// </summary>
+    public abstract class StackPanelBase : StackPanel
+    {
+    }
+}